[BUGFIX] Move description of command line options
Flow uses the source code comments to print a description
of all options of a command line action. The comments have
to be in the same order like the arguments passed to
the function.

This is currently not the case for the account:create command.

Rearange the comments to match the existing order
of the argument list.